<refactor> adjust file structure

+++ backend/init.sql
@@ -2,5 +2,5 @@ CREATE TABLE IF NOT EXISTS public.user (
                                            userid SERIAL PRIMARY KEY,
                                            username VARCHAR(50) NOT NULL,
                                            password VARCHAR(50) NOT NULL,
-                                           balance DECIMAL(10, 2) DEFAULT 0.00
+                                           balance INT DEFAULT 0
 );
